requirements.txt
numpy==1.22.3
pandas==1.4.2
python-dotenv==0.20.0
requests==2.27.1
streamlit==1.8.1
streamlit-lottie==0.0.3
toml==0.10.2
urllib3==1.26.9
XlsxWriter==3.0.3
zipp==3.8.0
